# my-auth-app/openapi.json
{"openapi":"3.0.3","info":{"title":"auth_api_rust","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/auth/login":{"post":{"tags":["handlers"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Invalid credentials"},"500":{"description":"Internal server error"}}}},"/api/auth/logout":{"post":{"tags":["handlers"],"operationId":"logout","responses":{"200":{"description":"Logout successful"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]}},"/api/auth/refresh":{"post":{"tags":["handlers"],"operationId":"refresh","responses":{"200":{"description":"Token refreshed successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Invalid or missing refresh token"},"500":{"description":"Internal server error"}},"security":[{"cookieAuth":[]}]}},"/api/auth/register":{"post":{"tags":["handlers"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"400":{"description":"Invalid input or email already exists"},"500":{"description":"Internal server error"}}}},"/api/users/profile":{"get":{"tags":["handlers"],"operationId":"profile","responses":{"200":{"description":"User profile data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"401":{"description":"Unauthorized"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"strongpassword123"}}},"LoginResponse":{"type":"object","required":["accessToken"],"properties":{"accessToken":{"type":"string"}}},"RegisterRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"strongpassword123"}}},"Role":{"type":"string","enum":["User","Admin"]},"User":{"type":"object","required":["id","email","role"],"properties":{"email":{"type":"string"},"id":{"type":"string","format":"uuid","example":"a1b2c3d4-e5f6-7890-1234-567890abcdef"},"role":{"$ref":"#/components/schemas/Role"}}},"UserProfileResponse":{"type":"object","required":["id","email","role"],"properties":{"email":{"type":"string"},"id":{"type":"string","format":"uuid","example":"a1b2c3d4-e5f6-7890-1234-567890abcdef"},"role":{"$ref":"#/components/schemas/Role"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"},"cookieAuth":{"type":"apiKey","in":"cookie","name":"jid"}}},"tags":[{"name":"Auth API","description":"Rust + Axum Authentication API"}]}